+++ services-core/src/test/java/com/mapbox/core/TestUtils.java
@@ -8,9 +8,11 @@ import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
+import java.io.InputStreamReader;
 import java.io.ObjectInputStream;
 import java.io.ObjectOutputStream;
 import java.io.Serializable;
+import java.nio.charset.StandardCharsets;
 import java.util.Scanner;
 
 import static java.nio.charset.StandardCharsets.UTF_8;
@@ -28,10 +30,18 @@ public class TestUtils {
   }
 
   protected String loadJsonFixture(String filename) throws IOException {
-    ClassLoader classLoader = getClass().getClassLoader();
-    InputStream inputStream = classLoader.getResourceAsStream(filename);
+    InputStream inputStream = getResourceInputSteam(filename);
     Scanner scanner = new Scanner(inputStream, UTF_8.name()).useDelimiter("\\A");
     return scanner.hasNext() ? scanner.next() : "";
+  }
+
+  protected InputStreamReader getResourceInputSteamReader(String filename) {
+    return new InputStreamReader(getResourceInputSteam(filename), StandardCharsets.UTF_8);
+  }
+
+  private InputStream getResourceInputSteam(String filename) {
+    ClassLoader classLoader = getClass().getClassLoader();
+    return classLoader.getResourceAsStream(filename);
   }
 
   public static <T extends Serializable> byte[] serialize(T obj)